Force having OOC notes set to spawn

+++ code/game/gamemodes/gameticker.dm
@@ -274,8 +274,8 @@ var/global/datum/controller/gameticker/ticker
 				else if(!player.mind.assigned_role)
 					continue
 				else
-					player.create_character()
-					qdel(player)
+					if (player.create_character()) // VOREStation Edit
+						qdel(player)
 
 
 	proc/collect_minds()
